+++ src/com/android/launcher3/widget/WidgetsRecyclerView.java
@@ -22,12 +22,12 @@ import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
 
-import com.android.launcher3.BaseRecyclerView;
-import com.android.launcher3.R;
-
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.recyclerview.widget.RecyclerView.OnItemTouchListener;
+
+import com.android.launcher3.BaseRecyclerView;
+import com.android.launcher3.R;
 
 /**
  * The widgets recycler view.
@@ -79,10 +79,10 @@ public class WidgetsRecyclerView extends BaseRecyclerView implements OnItemTouch
      * Maps the touch (from 0..1) to the adapter position that should be visible.
      */
     @Override
-    public String scrollToPositionAtProgress(float touchFraction) {
+    public PositionThumbInfo scrollToPositionAtProgress(float touchFraction) {
         // Skip early if widgets are not bound.
         if (isModelNotReady()) {
-            return "";
+            return new PositionThumbInfo("", 0);
         }
 
         // Stop the scroller if it is scrolling
@@ -95,7 +95,7 @@ public class WidgetsRecyclerView extends BaseRecyclerView implements OnItemTouch
         layoutManager.scrollToPositionWithOffset(0, (int) -(availableScrollHeight * touchFraction));
 
         int posInt = (int) ((touchFraction == 1) ? pos - 1 : pos);
-        return mAdapter.getSectionName(posInt);
+        return new PositionThumbInfo(mAdapter.getSectionName(posInt), 0);
     }
 
     /**