+++ res/raw/changelog.md
@@ -1,10 +1,11 @@
 ### v0.6.3
 *   Add Overview Mode for desktop
 *   Add web search providers
+*   Support for folders in appdrawer.
+*   support order by icon color
 
 ### v0.6.2
-*   Add Preferences for landscape and tables
-*   Enable global search in all aps
+*   Enable global search in all apps
 
 ### v0.6.1
 *   Support for android Pie